Underwater Tint removal and Damage Shake test
Screen Fades/Tints cannot be cancelled or over-written, at least to my current knowledge. The underwater tint was causing a serious issue with other tints in-game. This also applies to the death fade, causing it to not play if your screen is tinted, but that is a minor issue. I will continue to investigate a fix

On the upside, adding screenshakes to events like explosions, recoil, and damage. Testing on the zombies, it makes their attacks feel much more powerful!